Add two patterns for JDBC+MSSSQL error messages
... which I encountered during a pen test. A case of SQL Injection.
Thus it is labled as high. Firm indicates that it needs to be manually
confirmed. In this case it DID NOT work (confirmed with sqlmap and
somehow with the client).

I spotted them but not using burp-suite-error-message-checks

message was

{"msg":"\n### Error querying database.  Cause: com.microsoft.sqlserver.jdbc.SQLServerException: Invalid object name 'user'.\n### The error may exist in REDACTED.java (best guess)\n### The error may involve defaultParameterMap\n### The error occurred while setting parameters\n### SQL: SELECT  id,account,password,is_active,create_time,update_time,failed_numbers  FROM [user]     WHERE (account = ?)\n### Cause: com.microsoft.sqlserver.jdbc.SQLServerException: Invalid object name 'user'.\n; bad SQL grammar []; nested exception is com.microsoft.sqlserver.jdbc.SQLServerException: Invalid object name 'user'.","code":500}

fixes #66

The diff on two other lines I noticed seemed tohave different line endings. The don't
show with ``git diff -w`` .